Minor fixes
Adjusted the timing to 5 seconds for deleting DMs. Yea, it takes
longer, but until twitter decides that you can make more queries, then
you’ll have to suffer with slowing down queries.

Fixed an issue in the ssl cn grabber that if a CIDR value was entered,
it was stuck in the while loop. temp fix is to just look for the input
of “done” to be entered. It will then start to do it’s magic.