Migrate twitter menu
Moved the twitter menu options into the actual twitter lib. No reason
to clutter up the main script itself just to print a menu.

Added exception handling within the twitter lib so that if a user
accidentally enters a username into the integer field of the function
to gather X amount of tweets, the script doesn’t just crash.